requirements.txt
pytest
selenium
allure-pytest
pytest-rerunfailures
pytest-xdist
python-dotenv
webdriver_manager